+++ lib/src/requests_inspector_interceptor.dart
@@ -12,10 +12,27 @@ class RequestsInspectorInterceptor extends Interceptor {
         url: response.requestOptions.path,
         statusCode: response.statusCode ?? 0,
         headers: response.requestOptions.headers,
+        queryParameters: response.requestOptions.queryParameters,
         responseBody: response.data,
         sentTime: DateTime.now(),
       ),
     );
     super.onResponse(response, handler);
+  }
+
+  @override
+  void onError(DioError err, ErrorInterceptorHandler handler) {
+    InspectorController().addNewRequest(
+      RequestDetails(
+        requestMethod: RequestMethod.values
+            .firstWhere((e) => e.name == err.requestOptions.method),
+        url: err.requestOptions.path,
+        headers: err.requestOptions.headers,
+        queryParameters: err.requestOptions.queryParameters,
+        responseBody: err.message,
+        sentTime: DateTime.now(),
+      ),
+    );
+    super.onError(err, handler);
   }
 }